# Use NUM2LL() macro instead of rb_num2ll()