--- a/ext/oj/dump_compat.c
+++ b/ext/oj/dump_compat.c
@@ -476,12 +476,12 @@ time_alt(VALUE obj, int depth, Out out) {
 	sec = (long long)ts.tv_sec;
 	nsec = ts.tv_nsec;
     } else {
-	sec = rb_num2ll(rb_funcall2(obj, oj_tv_sec_id, 0, 0));
-	nsec = rb_num2ll(rb_funcall2(obj, oj_tv_nsec_id, 0, 0));
+	sec = NUM2LL(rb_funcall2(obj, oj_tv_sec_id, 0, 0));
+	nsec = NUM2LL(rb_funcall2(obj, oj_tv_nsec_id, 0, 0));
     }
 #else
-    sec = rb_num2ll(rb_funcall2(obj, oj_tv_sec_id, 0, 0));
-    nsec = rb_num2ll(rb_funcall2(obj, oj_tv_nsec_id, 0, 0));
+    sec = NUM2LL(rb_funcall2(obj, oj_tv_sec_id, 0, 0));
+    nsec = NUM2LL(rb_funcall2(obj, oj_tv_nsec_id, 0, 0));
 #endif
 
     attrs[0].num = sec;