--- a/ext/oj/odd.c
+++ b/ext/oj/odd.c
@@ -37,8 +37,8 @@ static VALUE get_datetime_secs(VALUE obj) {
     volatile VALUE rsecs = rb_funcall(obj, sec_id, 0);
     volatile VALUE rfrac = rb_funcall(obj, sec_fraction_id, 0);
     long           sec   = NUM2LONG(rsecs);
-    long long      num   = rb_num2ll(rb_funcall(rfrac, numerator_id, 0));
-    long long      den   = rb_num2ll(rb_funcall(rfrac, denominator_id, 0));
+    long long      num   = NUM2LL(rb_funcall(rfrac, numerator_id, 0));
+    long long      den   = NUM2LL(rb_funcall(rfrac, denominator_id, 0));
 
     num += sec * den;