By using NUM2LL() macro, seems that it can convert the value faster than rb_num2ll() directly.

```c
#define RB_NUM2LL  rb_num2ll_inline   /**< @alias{rb_num2ll_inline} */
#define NUM2LL     RB_NUM2LL          /**< @old{RB_NUM2LL} */

static inline LONG_LONG
rb_num2ll_inline(VALUE x)
{
    if (RB_FIXNUM_P(x))
        return RB_FIX2LONG(x);
    else
        return rb_num2ll(x);
}
```
https://github.com/ruby/ruby/blob/master/include/ruby/internal/arithmetic/long_long.h

−               | before | after  | result
--               | --     | --     | --
Oj.dump (macOS)  | 7.461k | 7.839k | 1.051x
Oj.dump (Linux)  | 8.391k | 9.140k | 1.089x

### Environment
- macOS
  - macOS 12.1
  - Apple M1 Max
  - Apple clang version 13.0.0 (clang-1300.0.29.30)
  - Ruby 3.1.0
- Linux
  - Zorin OS 16
  - AMD Ryzen 7 5700G
  - gcc version 11.1.0
  - Ruby 3.1.0

### macOS
#### Before
```
Warming up --------------------------------------
             Oj.dump   737.000  i/100ms
Calculating -------------------------------------
             Oj.dump      7.461k (± 1.1%) i/s -    112.024k in  15.015711s
```

#### After
```
Warming up --------------------------------------
             Oj.dump   783.000  i/100ms
Calculating -------------------------------------
             Oj.dump      7.839k (± 0.9%) i/s -    118.233k in  15.084888s
```

### Linux
#### Before
```
Warming up --------------------------------------
             Oj.dump   793.000  i/100ms
Calculating -------------------------------------
             Oj.dump      8.391k (± 0.9%) i/s -    126.087k in  15.028559s
```

#### After
```
Warming up --------------------------------------
             Oj.dump   887.000  i/100ms
Calculating -------------------------------------
             Oj.dump      9.140k (± 0.9%) i/s -    137.485k in  15.043569s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

data = (0..10000).to_a

Benchmark.ips do |x|
  x.time = 15

  x.report('Oj.dump') { Oj.dump(data, mode: :compat) }
end
```